update desktop/server scripts

+++ ubuntu/fuzzing_stack.sh
@@ -1,15 +1,19 @@
 #!/bin/bash -e
 
-sudo apt install -y gcc-13-plugin-dev
+if [[ ! -d  ~/.build/AFLplusplus ]]; then
 
-# llvm 17
-cd ~/.build
-yes '' | ~/.config/dot/ubuntu/llvm-update.sh 17 20
+    sudo apt install -y gcc-13-plugin-dev
 
-# afl++
-cd ~/.build
-rm -rf AFLplusplus
-git clone https://github.com/AFLplusplus/AFLplusplus
-cd AFLplusplus
-make source-only NO_NYX=1
-sudo make install
+    # llvm 17
+    cd ~/.build
+    yes '' | ~/.config/dot/ubuntu/llvm-update.sh 17 20
+
+    # afl++
+    cd ~/.build
+    rm -rf AFLplusplus
+    git clone https://github.com/AFLplusplus/AFLplusplus
+    cd AFLplusplus
+    make source-only NO_NYX=1
+    sudo make install
+
+fi
